Add split variants for 5 operations all split on variable i with split index 4
